client/setup folder now matches server/setup structure. moved client_rest to folder so that i could put it in a package to make running unittests against it easier
