+++ client/lib/client_rest.py
