+++ client/lib/test/image_raw_manual_test.py
@@ -1,0 +1,3 @@
+import unittest
+from client_rest import ImagingInterface
+from client_rest import ImageInfo
