+++ client/lib/test/interface_setup_test.py
@@ -1,0 +1,6 @@
+import unittest
+from client_rest import ImagingInterface
+
+class TestPing(unittest.TestCase):
+    def test(self):
+        print("heyyyyyyy")
