+++ client/setup/internal/pip-requirements.txt
